Convert MHC migration code to a pipeline job
