+++ primeseq/webapp/WEB-INF/primeseqContext.xml
@@ -3,6 +3,28 @@
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-2.0.xsd">
 
+    <bean id="primeseqPipelineRegistry" class="org.labkey.api.pipeline.TaskPipelineRegistrar">
+        <property name="factoryImpls">
+            <list>
+                <bean class="org.labkey.primeseq.pipeline.MhcMigrationPipelineJob.Task$Factory">
+                    <property name="location" value="webserver" />
+                </bean>
+            </list>
+        </property>
+        <property name="pipelines">
+            <list>
+                <bean class="org.labkey.api.pipeline.TaskPipelineSettings">
+                    <constructor-arg type="java.lang.Class" value="org.labkey.primeseq.pipeline.MhcMigrationPipelineJob"/>
+                    <property name="taskProgressionSpec">
+                        <list>
+                            <value type="java.lang.Class">org.labkey.primeseq.pipeline.MhcMigrationPipelineJob.Task</value>
+                        </list>
+                    </property>
+                </bean>
+            </list>
+        </property>
+    </bean>
+
     <!-- This is intended to force registration of the PipelineSteps -->
     <bean id="pipelineStartup" class="org.labkey.primeseq.PrimeseqModule$PipelineStartup">
